requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
alembic
pydantic>=2
pydantic-settings
python-multipart
celery
redis
requests
shapely
geopandas
geoalchemy2
pyproj
rasterio
numpy
Pillow
mercantile
matplotlib
